CleanCode-2 + code anh thựn

+++ src/main/java/com/VooTreeVeeVuu/domain/repository/HotelFacilityRepository.java
@@ -3,5 +3,6 @@ package com.VooTreeVeeVuu.domain.repository;
 import com.VooTreeVeeVuu.domain.entity.HotelFacility;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HotelFacilityRepository extends JpaRepository<HotelFacility, Integer> {
+public interface HotelFacilityRepository extends JpaRepository<HotelFacility, Long>{
+
 }
