+++ src/main/java/com/VooTreeVeeVuu/domain/repository/HotelImageRepository.java
@@ -3,5 +3,5 @@ package com.VooTreeVeeVuu.domain.repository;
 import com.VooTreeVeeVuu.domain.entity.HotelImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HotelImageRepository extends JpaRepository<HotelImage, Integer> {
+public interface HotelImageRepository extends JpaRepository<HotelImage, Long> {
 }
