+++ src/main/java/com/VooTreeVeeVuu/domain/repository/LogsRepository.java
@@ -3,5 +3,5 @@ package com.VooTreeVeeVuu.domain.repository;
 import com.VooTreeVeeVuu.domain.entity.Logs;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LogsRepository extends JpaRepository<Logs, Integer> {
+public interface LogsRepository extends JpaRepository<Logs, Long> {
 }
