+++ src/main/java/com/VooTreeVeeVuu/domain/repository/RoomFacilityRepository.java
@@ -3,5 +3,5 @@ package com.VooTreeVeeVuu.domain.repository;
 import com.VooTreeVeeVuu.domain.entity.RoomFacility;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoomFacilityRepository extends JpaRepository<RoomFacility, Integer> {
+public interface RoomFacilityRepository extends JpaRepository<RoomFacility, Long> {
 }
